feat: 조건부 렌더링 추가수정 에러수정
