+++ src/app/(layout)/page.tsx
@@ -1,41 +1,11 @@
 'use client';
-import Header from '@/components/MainPage/Header';
-import Main from '@/components/MainPage/Main';
-import useAuthStore from '@/zustand/useAuthStore';
-import { useEffect, useState } from 'react';
-import WebMainPage from '../(webmain)/webmain/page';
+import dynamic from 'next/dynamic';
+
+const HomeMain = dynamic(() => import('@/components/HomeMain/HomeMain'), {
+  ssr: false,
+});
+
 function Home() {
-  const { user } = useAuthStore();
-  const [isMobile, setIsMobile] = useState<boolean>(Boolean);
-
-  useEffect(() => {
-    // 윈도우 창의 크기를 감지하는 함수
-    const handleResize = () => {
-      setIsMobile(window.innerWidth < 1000);
-    };
-
-    // 컴포넌트가 처음 마운트 될 때와 창 크기 변화 시에 실행
-    handleResize(); // 초기 실행
-    window.addEventListener('resize', handleResize);
-
-    // 컴포넌트가 언마운트 될 때 이벤트 리스너 제거
-    return () => window.removeEventListener('resize', handleResize);
-  }, [isMobile]);
-
-  return (
-    <>
-      {window.innerWidth < 1000 ? (
-        <>
-          <Header
-            title={user ? `${user.nickname}님만의` : '여러분만의'}
-            content="맛집, 여행지를 공유해주세요!"
-          />
-          <Main />
-        </>
-      ) : (
-        <WebMainPage />
-      )}
-    </>
-  );
+  return <HomeMain />;
 }
 export default Home;
