+++ src/components/FeedList/UserName.tsx
@@ -4,9 +4,7 @@ import React from 'react';
 
 function UserName() {
   const { user } = useAuthStore();
-  return (
-    <div>{user ? `${user.user_metadata.nickname}님만의` : '여러분만의'}</div>
-  );
+  return <div>{user ? `${user.nickname}님만의` : '여러분만의'}</div>;
 }
 
 export default UserName;
