+++ src/components/MainPage/Main/Sections/WebFeedSection.tsx
@@ -18,6 +18,7 @@ function WebFeedSection() {
     const res = await axios.get('/api/supabase-feeduserdata');
     const items = res.data.data as TableFeedUserType[];
     setFeedItems(items);
+    console.log(items);
     setIsLoading(false);
   };
